refactors mock data logic
