+++ api/routes/book-related.js
@@ -1,5 +1,5 @@
 import { Hono } from "hono";
-import { handleDataFetching } from "../lib/utils.js";
+import { selectDataSource, bookRelatedMockUtils } from "../lib/utils.js";
 
 // Create book related router
 const bookRelatedRouter = new Hono();
@@ -8,46 +8,9 @@ const bookRelatedRouter = new Hono();
 bookRelatedRouter.get("/", async (c) => {
   const bookId = c.req.param("id");
   
-  // Mock data logic
+  // Use the imported mock logic
   const mockLogic = async (c) => {
-    const bookIdNum = parseInt(bookId, 10);
-    const book = c.env.MOCK_DATA.find(book => book.id === bookIdNum);
-    
-    if (!book) {
-      return Response.json({ error: "Book not found" }, { status: 404 });
-    }
-
-    const bookGenre = book.genre;
-    
-    // Generate mock related data
-    const relatedBooks = c.env.MOCK_DATA
-      .filter(b => b.genre === bookGenre && b.id !== bookIdNum)
-      .slice(0, 3);
-    
-    // Generate mock recent books
-    const recentBooks = c.env.MOCK_DATA
-      .filter(b => b.id !== bookIdNum)
-      .slice(0, 2);
-    
-    // Generate mock genre counts
-    const genres = {};
-    c.env.MOCK_DATA.forEach(b => {
-      genres[b.genre] = (genres[b.genre] || 0) + 1;
-    });
-    
-    const genreCounts = Object.entries(genres).map(([genre, count]) => ({
-      genre,
-      count
-    })).sort((a, b) => b.count - a.count);
-    
-    return Response.json({
-      bookId: bookId,
-      bookGenre: bookGenre,
-      relatedBooks,
-      recentRecommendations: recentBooks,
-      genreStats: genreCounts,
-      source: "mock"
-    });
+    return bookRelatedMockUtils.getRelatedBookData(c, bookId);
   };
   
   // Database logic
@@ -93,7 +56,7 @@ bookRelatedRouter.get("/", async (c) => {
     });
   };
   
-  return handleDataFetching(c, dbLogic, mockLogic);
+  return selectDataSource(c, dbLogic, mockLogic);
 });
 
 export default bookRelatedRouter;
